% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgca_package.R
\docType{data}
\name{ages_darmanis}
\alias{ages_darmanis}
\title{Brain sample ages vector.}
\format{An object of class \code{numeric} of length 158.}
\usage{
ages_darmanis
}
\description{
A vector specifying the ages of the brain cell types from the single-cell RNA-seq study. The total data set can be downloaded by following the links in the original paper.
}
\references{
Darmanis S, Sloan SA, Zhang Y, et al. A survey of human brain transcriptome diversity at the single cell level. Proc Natl Acad Sci USA. 2015;112(23):7285-90.
}
\keyword{datasets}

