% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddcorClasses.R
\docType{class}
\name{corMats-class}
\alias{corMats-class}
\title{An S4 class to store correlation matrices and associated info.}
\description{
An S4 class to store correlation matrices and associated info.
}
\section{Slots}{

\describe{
\item{\code{corMatList}}{List of correlation, number of sample, and correlation significance matrices in each condition.}

\item{\code{design}}{Design matrix inputted into the function.}

\item{\code{options}}{Character vector of options used in the function.}
}}

