% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgca_package.R
\docType{data}
\name{darmanis}
\alias{darmanis}
\title{Single-cell gene expression data from different brain cell types.}
\format{An object of class \code{data.frame} with 572 rows and 158 columns.}
\usage{
darmanis
}
\description{
This data set has been filtered to include the genes that are in the 95th percentile or above in both oligodendrocytes and neurons. The total data set can be downloaded by following the links in the original paper.
}
\references{
Darmanis S, Sloan SA, Zhang Y, et al. A survey of human brain transcriptome diversity at the single cell level. Proc Natl Acad Sci USA. 2015;112(23):7285-90.
}
\keyword{datasets}

