% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddcorClasses.R
\docType{class}
\name{dcPair-class}
\alias{dcPair-class}
\title{S4 class for pairwise differential correlation matrices and associated info.}
\description{
S4 class for pairwise differential correlation matrices and associated info.
}
\section{Slots}{

\describe{
\item{\code{corA}}{Correlation matrix for identifiers in condition A.}

\item{\code{corPvalA}}{Matrix of correlation significances in condition A.}

\item{\code{corB}}{Correlation matrix for identifiers in condition B.}

\item{\code{corPvalB}}{Matrix of correlation significances in condition B.}

\item{\code{ZDiff}}{Matrix of differences in z-values between conditions.}

\item{\code{PValDiff}}{Matrix of p-values for differences in z-values between conditions.}

\item{\code{options}}{Character vector of options used in the function.}
}}

