% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgca_package.R
\docType{data}
\name{design_mat}
\alias{design_mat}
\title{Design matrix of cell type specifications of the single-cell RNA-seq samples.}
\format{An object of class \code{matrix} with 158 rows and 2 columns.}
\usage{
design_mat
}
\description{
Cell type specifications were performed by the authors. The total data set can be downloaded by following the links in the original paper.
}
\references{
Darmanis S, Sloan SA, Zhang Y, et al. A survey of human brain transcriptome diversity at the single cell level. Proc Natl Acad Sci USA. 2015;112(23):7285-90.
}
\keyword{datasets}

