% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgca_package.R
\docType{package}
\name{DGCA}
\alias{DGCA}
\alias{DGCA-package}
\title{DGCA: An R package for Differential Gene Correlation Analysis}
\description{
The DGCA package provides three major classes of functions:
1) Functions to calculate correlations, correlation significances, and number of samples in each correlation calculation based on an input matrix and a design matrix.
2) Functions to calculate differential correlations between regions, which in the current package can only be pairwise (i.e., one condition vs another).
3) Functions to extract, sort information about the differential correlation calculations in a convenient format.
The first two functions comprise the discovery of differential correlation (ddcor) portion of the package, which is why the names of the functions and object names often begin with ddcor.
Note that DGCA makes use of the 	SAF = getOption("stringsAsFactors"); on.exit(options(stringsAsFactors = SAF)); options(stringsAsFactors = FALSE) design pattern many times in order to avoid errors related to stringsAsFactors in porting code to new environments. This should not affect the stringsAsFactors options in your environment; however, you may want to be aware of this.
}

