% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleGO.R
\name{extractModuleGO}
\alias{extractModuleGO}
\title{Extract results from the module GO analysis}
\usage{
extractModuleGO(moduleGO_list, labels = NULL)
}
\arguments{
\item{moduleGO_list}{The list of list of data frames for each module to be turned into a data frame.}

\item{labels}{Optional, a list of module names. Optional; if not inputted, these will be extracted from the list of module GO enrichments.}
}
\value{
A data frame summarizing the GO term enrichments from each group, with columns on the ordered by the minimum p-value for OR term enrichment in any group.
}
\description{
Turns the list of lists from the moduleGO function into a more comprehensible data frame. Note that if a GO term enrichment does not exist for that module, it is set as NA.
}

