% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matCorSig.R
\name{matCorSig.fuzz}
\alias{matCorSig.fuzz}
\title{Calculate correlation matrix p-values, adding a small noise term to the result to ensure normality.}
\usage{
matCorSig.fuzz(corrs, nsamp, secondMat = FALSE, epsilon = 1e-06)
}
\arguments{
\item{corrs}{Computed correlation matrix.}

\item{nsamp}{Computed number of samples used per call in the correlation matrix.}

\item{secondMat}{Logical indicator of whether there is a second matrix in the comparison or not.}

\item{epsilon}{A number representing the maximum noise to add to each p-value. Default is 1e-6.}
}
\value{
A matrix of p-values.
}
\description{
Calculate two-sided p-values from a pairwise correlations matrix and a corresponding "number of samples" matrix, adding a small random noise term which is defined as epsilon * rand(0,1).
}
\references{
HMisc R package https://cran.r-project.org/web/packages/Hmisc/index.html; Yang et al., A Fuzzy Permutation Method for False Discovery Rate Control, Nat Sci Rep (2016), https://www.nature.com/articles/srep28507
}
