% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matCorSig.R
\name{matCorSig}
\alias{matCorSig}
\title{Calculate correlation matrix p-values}
\usage{
matCorSig(corrs, nsamp, secondMat = FALSE)
}
\arguments{
\item{corrs}{Computed correlation matrix.}

\item{nsamp}{Computed number of samples used per call in the correlation matrix.}

\item{secondMat}{Logical indicator of whether there is a second matrix in the comparison or not.}
}
\value{
A matrix of p-values.
}
\description{
Calculate two-sided p-values from a pairwise correlations matrix and a corresponding "number of samples" matrix.
}
\references{
HMisc R package https://cran.r-project.org/web/packages/Hmisc/index.html
}

