% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matNSamp.R
\name{matNSamp}
\alias{matNSamp}
\title{Find the number of non-missing values.}
\usage{
matNSamp(matA, impute = FALSE, matB = NULL, secondMat = FALSE)
}
\arguments{
\item{matA}{Input data matrix with numeric entries.}

\item{impute}{Binary value; if true, indicates that imputation was performed previously, and so checking for NAs is not necessary.}

\item{matB}{Optional input data matrix with which the comparison with matA will be made.}

\item{secondMat}{Logical indicator of whether there is a second matrix in the comparison or not.}
}
\value{
A number of samples (nsamp) matrix.
}
\description{
This function calculates the pairwise number of non-missing values in a matrix.
}

