% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddcorGO.R
\name{switchGenesToHGCN}
\alias{switchGenesToHGCN}
\title{Switches a gene vector to cleaned HGNC symbols}
\usage{
switchGenesToHGCN(gene_list)
}
\arguments{
\item{gene_list}{Character vector of gene names.}
}
\value{
Character vector of cleaned gene names.
}
\description{
Where possible, switches a character vector of gene names to cleaned and updated HGNC symbols
}

